My first commit

+++ 0x0C-more_malloc_free/101-mul.c
@@ -8,14 +8,14 @@
  */
 int main(int argc, char *argv[])
 {
-	unsigned int res1, res2;
+	unsigned long int res1, res2;
 
 	res1 = res2 = 0;
 	if (argc == 3)
 	{
 		res1 = atoi(argv[1]);
 		res2 = atoi(argv[2]);
-		printf("%u\n", res1 * res2);
+		printf("%lu\n", res1 * res2);
 		return (0);
 	}
 	else
